requirements.txt
# Function dependencies, for example:
# package>=version 
azure-core==1.28.0
azure-ai-formrecognizer==3.2.0
google-cloud-storage==2.9.0
pandas>=1.3.3
pandas-gbq>=0.13.2
paramiko==2.7.2
requests==2.26.0 
Flask==2.1.2
Werkzeug==2.1.2
google-cloud-secret-manager==2.7.0
google-api-python-client==2.33.0
functions-framework
google-cloud-pubsub==2.15.0
google-api-python-client==2.33.0
google-cloud-bigquery-storage>=2.21.0